add more strict auth-ing

+++ src/routes/events.rs
@@ -13,23 +13,34 @@ use axum::{
     extract::{Query, State},
 };
 use maud::{Markup, html};
+use crate::auth::{AuthUtilities, PermissionsTarget};
 
 #[axum::debug_handler]
 pub async fn get_events(State(state): State<DenimState>, session: DenimSession) -> Markup {
+    let can_add_events = session.can(PermissionsTarget::CRUD_EVENTS);
+    
     state.render(session, html!{
         div class="mx-auto bg-gray-800 p-8 rounded shadow-md max-w-4xl w-full flex flex-col space-y-4" {
             div hx-ext="sse" sse-connect="/sse_feed" class="container flex flex-row justify-center space-x-4" {
                 div hx-get="/internal/get_events" hx-trigger="sse:crud_event,load" id="all_events" {}
-                div id="in_focus" hx-get="/internal/get_events_form" hx-trigger="load" {}
+                @if can_add_events {
+                    div id="in_focus" hx-get="/internal/get_events_form" hx-trigger="load" {}
+                } @else {
+                   div id="in_focus" {} 
+                }
             }
-            button class="bg-blue-600 hover:bg-blue-800 font-bold py-2 px-4 rounded" hx-get="/internal/get_events_form" hx-target="#in_focus" {
-                "Add new Event"
-            }   
+            @if can_add_events {
+                button class="bg-blue-600 hover:bg-blue-800 font-bold py-2 px-4 rounded" hx-get="/internal/get_events_form" hx-target="#in_focus" {
+                    "Add new Event"
+                }   
+            }
         }
     })
 }
 
-pub async fn internal_get_add_events_form(State(state): State<DenimState>) -> DenimResult<Markup> {
+pub async fn internal_get_add_events_form(State(state): State<DenimState>, session: DenimSession) -> DenimResult<Markup> {
+    session.ensure_can(PermissionsTarget::CRUD_EVENTS)?;
+    
     let staff = User::get_all_staff(&state).await?;
 
     Ok(html! {
@@ -55,6 +66,7 @@ pub async fn internal_get_add_events_form(State(state): State<DenimState>) -> De
 
 pub async fn put_new_event(
     State(state): State<DenimState>,
+    session: DenimSession,
     Form(add_event_form): Form<<Event as DataType>::FormForAdding>,
 ) -> DenimResult<Markup> {
     let id =
@@ -62,7 +74,8 @@ pub async fn put_new_event(
     state.send_sse_event(SseEvent::CrudEvent);
 
     let this_event =
-        internal_get_event_in_detail(State(state.clone()), Query(IdForm { id })).await?;
+        internal_get_event_in_detail(State(state.clone()), session, Query(IdForm { id })).await?;
+    
     Ok(html! {
         (this_event)
     })
@@ -70,12 +83,15 @@ pub async fn put_new_event(
 
 pub async fn delete_event(
     State(state): State<DenimState>,
+    session: DenimSession,
     Query(IdForm { id }): Query<IdForm>,
 ) -> DenimResult<Markup> {
+    session.ensure_can(PermissionsTarget::CRUD_EVENTS)?;
+    
     Event::remove_from_database(id, &mut *state.get_connection().await?).await?;
     state.send_sse_event(SseEvent::CrudEvent);
 
-    let form = internal_get_add_events_form(State(state.clone())).await?;
+    let form = internal_get_add_events_form(State(state.clone()), session).await?;
     Ok(html! {
         (form)
     })
@@ -83,12 +99,16 @@ pub async fn delete_event(
 
 pub async fn internal_get_event_in_detail(
     State(state): State<DenimState>,
+    session: DenimSession,
     Query(IdForm { id }): Query<IdForm>,
 ) -> DenimResult<Markup> {
     let Some(event) = Event::get_from_db_by_id(id, &mut *state.get_connection().await?).await?
     else {
         return Err(DenimError::MissingEvent { id });
     };
+    
+    let can_view_sensitives = session.can(PermissionsTarget::VIEW_SENSITIVE_DETAILS);
+    let can_delete = session.can(PermissionsTarget::CRUD_EVENTS);
 
     Ok(html! {
         div hx-get="/internal/get_event" hx-target="#in_focus" hx-vals={"{\"id\": \"" (id) "\"}" } hx-trigger="sse:crud_event" {
@@ -105,10 +125,12 @@ pub async fn internal_get_event_in_detail(
                     "Time: "
                     span class="font-medium" {(event.date.format("%a %d/%m/%y @ %H:%M"))}
                 }
-                @if let Some(staff) = event.associated_staff_member {
-                    p class="text-gray-200 font-semibold" {
-                        "Staff Member: "
-                        span class="font-medium" {(staff)}
+                @if can_view_sensitives {
+                    @if let Some(staff) = event.associated_staff_member {
+                        p class="text-gray-200 font-semibold" {
+                            "Staff Member: "
+                            span class="font-medium" {(staff)}
+                        }
                     }
                 }
                 @if let Some(extra) = event.extra_info {
@@ -117,9 +139,11 @@ pub async fn internal_get_event_in_detail(
                         span class="font-medium" {(extra)}
                     }
                 }
-                br;
-                button class="bg-red-600 hover:bg-red-800 font-bold py-2 px-4 rounded" hx-delete="/events" hx-vals={"{\"id\": \"" (id) "\"}" } hx-target="#in_focus" {
-                    "Delete event"
+                @if can_delete {
+                    br;
+                    button class="bg-red-600 hover:bg-red-800 font-bold py-2 px-4 rounded" hx-delete="/events" hx-vals={"{\"id\": \"" (id) "\"}" } hx-target="#in_focus" {
+                        "Delete event"
+                    }
                 }
             }
         }
